refactor(agents): move checkVersionCompatibility to BaseAgentAdapter
The method was duplicated across ClaudePlugin and GeminiPlugin with
identical comparison logic. Moved it to BaseAgentAdapter so all agents
share a single implementation — agents only need to override getVersion()
for agent-specific output parsing.

- Add checkVersionCompatibility() to BaseAgentAdapter using
  metadata.supportedVersion and metadata.minimumSupportedVersion
- Simplify run() version-check guard: replace duck-type
  'checkVersionCompatibility in this' check with direct
  this.metadata.supportedVersion presence check
- Remove duplicate checkVersionCompatibility() from ClaudePlugin and
  GeminiPlugin
- Clean up now-unused compareVersions/createErrorContext/
  VersionCompatibilityResult imports in both plugin files
