+++ src/agents/plugins/gemini/gemini.plugin.ts
@@ -1,13 +1,10 @@
-import type { AgentMetadata, HookTransformer, VersionCompatibilityResult } from '../../core/types.js';
+import type { AgentMetadata, HookTransformer } from '../../core/types.js';
 import { BaseAgentAdapter } from '../../core/BaseAgentAdapter.js';
 import { GeminiHookTransformer } from './gemini.hook-transformer.js';
 import { GeminiSessionAdapter } from './gemini.session-adapter.js';
 import type { SessionAdapter } from '../../core/session/BaseSessionAdapter.js';
 import { GeminiExtensionInstaller } from './gemini.extension-installer.js';
 import type { BaseExtensionInstaller } from '../../core/extension/BaseExtensionInstaller.js';
-import { compareVersions } from '../../../utils/version-utils.js';
-import { createErrorContext } from '../../../utils/errors.js';
-import { logger } from '../../../utils/logger.js';
 
 /**
  * Supported Gemini CLI version
@@ -227,112 +224,4 @@ export class GeminiPlugin extends BaseAgentAdapter {
     }
   }
 
-  /**
-   * Check if installed version is compatible with CodeMie
-   * Compares against metadata.supportedVersion and metadata.minimumSupportedVersion
-   *
-   * @returns Version compatibility result with status and version info
-   */
-  async checkVersionCompatibility(): Promise<VersionCompatibilityResult> {
-    const supportedVersion = this.metadata.supportedVersion || 'latest';
-    const minimumSupportedVersion = this.metadata.minimumSupportedVersion;
-
-    const installedVersion = await this.getVersion();
-
-    logger.debug('Checking version compatibility', {
-      agent: this.metadata.name,
-      installedVersion,
-      supportedVersion,
-      minimumSupportedVersion,
-    });
-
-    if (!installedVersion) {
-      return {
-        compatible: false,
-        installedVersion: null,
-        supportedVersion,
-        isNewer: false,
-        hasUpdate: false,
-        isBelowMinimum: false,
-        minimumSupportedVersion,
-      };
-    }
-
-    if (!this.metadata.supportedVersion) {
-      return {
-        compatible: true,
-        installedVersion,
-        supportedVersion: 'latest',
-        isNewer: false,
-        hasUpdate: false,
-        isBelowMinimum: false,
-        minimumSupportedVersion,
-      };
-    }
-
-    try {
-      const comparison = compareVersions(installedVersion, supportedVersion);
-      const hasUpdate = comparison < 0;
-
-      let isBelowMinimum = false;
-      if (minimumSupportedVersion) {
-        const minimumComparison = compareVersions(installedVersion, minimumSupportedVersion);
-        isBelowMinimum = minimumComparison < 0;
-      }
-
-      logger.debug('Version comparison result', {
-        agent: this.metadata.name,
-        comparison,
-        installedVersion,
-        supportedVersion,
-        minimumSupportedVersion,
-        compatible: comparison <= 0,
-        isNewer: comparison > 0,
-        hasUpdate,
-        isBelowMinimum,
-      });
-
-      return {
-        compatible: comparison <= 0,
-        installedVersion,
-        supportedVersion,
-        isNewer: comparison > 0,
-        hasUpdate,
-        isBelowMinimum,
-        minimumSupportedVersion,
-      };
-    } catch (error) {
-      const errorContext = createErrorContext(error, { agent: this.metadata.name });
-      const isParseError =
-        error instanceof Error && error.message.includes('Invalid semantic version');
-
-      if (isParseError) {
-        logger.warn('Non-standard version format detected, treating as incompatible', {
-          ...errorContext,
-          operation: 'checkVersionCompatibility',
-          installedVersion,
-          supportedVersion,
-          minimumSupportedVersion,
-        });
-      } else {
-        logger.error('Version compatibility check failed unexpectedly', {
-          ...errorContext,
-          operation: 'checkVersionCompatibility',
-          installedVersion,
-          supportedVersion,
-          minimumSupportedVersion,
-        });
-      }
-
-      return {
-        compatible: false,
-        installedVersion,
-        supportedVersion,
-        isNewer: false,
-        hasUpdate: false,
-        isBelowMinimum: false,
-        minimumSupportedVersion,
-      };
-    }
-  }
 }
